chore(web): split api.ts god-file + dedupe formatters + prune dead UI/i18n
