--- web/src/components/ui/direction.tsx
@@ -1,5 +1,0 @@
-// metapi-go/ui — direction component ported from newapi (base-nova style, @base-ui/react). AGPL header stripped.
-export {
-  DirectionProvider,
-  useDirection,
-} from '@base-ui/react/direction-provider'
