--- web/src/components/ui/kbd.tsx
@@ -1,27 +1,0 @@
-// metapi-go/ui — kbd component ported from newapi (base-nova style, @base-ui/react). AGPL header stripped.
-import { cn } from '@/lib/utils'
-
-function Kbd({ className, ...props }: React.ComponentProps<'kbd'>) {
-  return (
-    <kbd
-      data-slot='kbd'
-      className={cn(
-        "bg-muted text-muted-foreground in-data-[slot=tooltip-content]:bg-background/20 in-data-[slot=tooltip-content]:text-background dark:in-data-[slot=tooltip-content]:bg-background/10 pointer-events-none inline-flex h-5 w-fit min-w-5 items-center justify-center gap-1 rounded-sm px-1 font-mono text-xs font-medium select-none [&_svg:not([class*='size-'])]:size-3",
-        className
-      )}
-      {...props}
-    />
-  )
-}
-
-function KbdGroup({ className, ...props }: React.ComponentProps<'div'>) {
-  return (
-    <kbd
-      data-slot='kbd-group'
-      className={cn('inline-flex items-center gap-1', className)}
-      {...props}
-    />
-  )
-}
-
-export { Kbd, KbdGroup }
