+++ web/src/features/accounts/api.ts
@@ -19,6 +19,7 @@ import {
 
 import i18n from '@/i18n/config'
 import { api } from '@/lib/api'
+import { assertBusinessOk } from '@/lib/assert-business-ok'
 import { toast } from '@/lib/toast'
 
 import type { AccountPayload, AccountStatus, AccountsSnapshot } from './types'
@@ -31,26 +32,6 @@ export const accountQueryKeys = {
   all: ['accounts'] as const,
   snapshot: () => [...accountQueryKeys.all, 'snapshot'] as const,
   detail: (id: number) => ['accounts', 'detail', id] as const,
-}
-
-// ---------------------------------------------------------------------------
-// Envelope helper — backend returns {success, message, data} for writes.
-// http-client already toasted the failure; we only throw to flip the mutation
-// to its error state and reject mutateAsync.
-// ---------------------------------------------------------------------------
-
-function assertBusinessOk<T>(result: unknown, fallback: string): T {
-  const envelope = result as {
-    success?: unknown
-    message?: unknown
-    data?: unknown
-  }
-  if (envelope && typeof envelope.success === 'boolean' && !envelope.success) {
-    throw new Error(
-      typeof envelope.message === 'string' ? envelope.message : i18n.t(fallback)
-    )
-  }
-  return (result as T) ?? (envelope?.data as T)
 }
 
 // ---------------------------------------------------------------------------
