+++ web/src/features/token-routes/api.ts
@@ -13,6 +13,7 @@ import {
 
 import i18n from '@/i18n/config'
 import { api } from '@/lib/api'
+import { assertBusinessOk } from '@/lib/assert-business-ok'
 import {
   normalizeMissingTokenModels,
   type MissingTokenModelsByName,
@@ -36,24 +37,6 @@ export const routeQueryKeys = {
   summary: () => [...routeQueryKeys.all, 'summary'] as const,
   candidates: () => [...routeQueryKeys.all, 'candidates'] as const,
   channels: (id: number) => ['routes', 'channels', id] as const,
-}
-
-// ---------------------------------------------------------------------------
-// Envelope helper
-// ---------------------------------------------------------------------------
-
-function assertBusinessOk<T>(result: unknown, fallback: string): T {
-  const envelope = result as {
-    success?: unknown
-    message?: unknown
-    data?: unknown
-  }
-  if (envelope && typeof envelope.success === 'boolean' && !envelope.success) {
-    throw new Error(
-      typeof envelope.message === 'string' ? envelope.message : i18n.t(fallback)
-    )
-  }
-  return (result as T) ?? (envelope?.data as T)
 }
 
 // ---------------------------------------------------------------------------
