--- web/src/hooks/index.ts
@@ -1,4 +1,0 @@
-// metapi-go/hooks — barrel re-exports. select.tsx imports useMediaQuery from '@/hooks'.
-// Add new hooks here as they land.
-
-export { useMediaQuery } from './use-media-query'
