+++ web/src/lib/api/search.ts
@@ -1,0 +1,10 @@
+import { request } from './transport'
+
+export const searchApi = {
+  // Search
+  search: (query: string) =>
+    request('/api/search', {
+      method: 'POST',
+      body: JSON.stringify({ query, limit: 20 }),
+    }),
+}
